requirements.txt
anthropic>=0.40.0
feedparser>=6.0.10
httpx>=0.27.0
python-dotenv>=1.0.0
